requirements.txt
django-extensions==0.9.3